Updates to go1.5.2 and includes protoc.
One big change is that go 1.5 requires go 1.4 to compile, which I didn't
want to deal with. It turns out that the binaries from the official
sources include the stdlib source code that you want for local godoc and
reference, so the compilation was not a useful step anyway. Also updated
to my major ubuntu upgrade and new user!